+++ hpc/nvidia/cuda/fakeCUDA/cuda-module-mgmt-luca.cpp
@@ -1,0 +1,2 @@
+#define __USE_LUCA__
+#include "cuda-module-mgmt.cpp"