+++ hpc/nvidia/cuda/fakeCUDA/runtime_event-luca.cpp
@@ -1,0 +1,2 @@
+#define __USE_LUCA__
+#include "runtime_event.cpp"